refactor(agent): keep compression changes within lint budgets
Problem:
- The explicit compression implementation pushed range planning over the complexity and function-size limits.
- New positional and boot coverage extended already oversized test files.

Fix:
- Split range planning, preparation, projection building, and atomic installation into focused helpers.
- Reuse the common compaction telemetry constructor.
- Move positional compression and allowlist coverage into dedicated test files.
- Keep the context compressor binding and UI copy changes line-neutral in baselined files.

Verification:
- go run ./tools/repolint
- go test ./...
- go test -race ./internal/agent ./internal/control
- go vet ./...
- cd desktop && go test ./...
- cd desktop/frontend && pnpm typecheck
