+++ internal/boot/boot_test.go
@@ -2183,8 +2183,7 @@ func defaultFullBootToolNames() []string {
 		"bash",
 		"bash_output",
 		"code_index",
-		"complete_step",
-		"compress",
+		"complete_step", "compress",
 		"delete_range",
 		"delete_symbol",
 		"docs",
@@ -2234,8 +2233,7 @@ func economyBootToolNames() []string {
 	return []string{
 		"ask",
 		"bash",
-		"bash_output",
-		"compress",
+		"bash_output", "compress",
 		"connect_tool_source",
 		"edit_file",
 		"kill_shell",
@@ -2287,8 +2285,7 @@ command = "reasonix-missing-mockmcp"
 	wantTools := []string{
 		"ask",
 		"bash",
-		"bash_output",
-		"compress",
+		"bash_output", "compress",
 		"connect_tool_source",
 		"edit_file",
 		"kill_shell",
@@ -2325,15 +2322,6 @@ command = "reasonix-missing-mockmcp"
 	}
 	if strings.Contains(sys, "# Skills") || strings.Contains(sys, "projskill") {
 		t.Fatalf("skills index should not be in economy system prompt:\n%s", sys)
-	}
-}
-
-func TestTokenEconomyCompressHonorsExplicitAllowlist(t *testing.T) {
-	if got := tokenEconomyBuiltins([]string{"read_file"}); slices.Contains(got, "compress") {
-		t.Fatalf("explicit allowlist unexpectedly enabled compress: %v", got)
-	}
-	if got := tokenEconomyBuiltins([]string{"compress"}); !reflect.DeepEqual(got, []string{"compress"}) {
-		t.Fatalf("explicit compress allowlist = %v, want [compress]", got)
 	}
 }
